+++ NetworkAdmin.App/SessionReportWindow.xaml.cs
@@ -1,0 +1,10 @@
+using System.Windows;
+
+namespace NetworkAdmin.App;
+
+public partial class SessionReportWindow : Window
+{
+    public SessionReportWindow() => InitializeComponent();
+
+    public void ShowReport(string html) => ReportBrowser.NavigateToString(html);
+}